Adds buttons (non-functional) to Question Screen. Minor Refactorization.
