+++ frontend/faculty_system/src/Components/HomeBox/ContentBox/InnerBoxes/EntitiesBox/EntitiesBox.js
@@ -1,14 +1,14 @@
 import React, {useState} from 'react';
-import SelectionBox from './SelectionBox';
+import SelectionBox from '../SelectionBox';
 
-export function ContactsBox() {
+export function EntitiesBox() {
 
     const [selected, setSelected] = useState("");
 
     return (
         <>
             <div id="selection-wrapper">
-                <SelectionBox name="contacts" content={[{name:"Contact 1"}, {name:"Contact 2"}]} update={setSelected} />
+                <SelectionBox name="entities" content={[{name:"Entity 1"}, {name:"Entity 2"}]} update={setSelected} />
             </div>
             <div id="content">
