+++ frontend/faculty_system/src/Components/HomeBox/ContentBox/InnerBoxes/QuestionsBox/questionsbox.css
@@ -1,20 +1,21 @@
 #new-question-selection {
-    height: 10%;
+    height: 12%;
     display: flex;
     align-items: center;
     text-align: center;
 }
 
 #selection-header {
-    height:8%;
+    height:10%;
     display:flex;
     align-items: center;
+    justify-content: center;
     border-bottom: 1px solid black;
 }
 
 #question-content {
     display:flex;
-    height: 100%;
+    height: 92%;
 }
 
 #response-box {
@@ -28,10 +29,18 @@
     height:100%;
 }
 
+#entities {
+    display: flex;
+    margin-left:4%;
+    margin-right:4%;
+}
 
 .section-title {
     height: 100%;
     width:25%;
+    display: flex; 
+    align-items: center; 
+    justify-content: center; 
     font-weight: bold;
     font-size: xx-large;
     border-right: 1px solid black;
@@ -92,4 +101,17 @@
     cursor: pointer;
     font-weight: bold;
     font-size: large;
+}
+.entity-selection-box {
+    flex:1;
+    margin: 2%;
+    display:flex;
+    align-items: center; 
+    justify-content: center; 
+    flex-direction: column; 
+    align-items: center; 
+}
+
+.entity-select {
+    width:100%;
 }
